Namespace prefix support.
A UCD word can now have a namespace prefix ; if it has none, it will be
considered as the default one: "ivoa".

The default policy to search UCD words has not change: namespace are ignored.
This is the default behaviour of UCDWordList.get(String), but it is possible
to take into account namespaces using UCDWordList.get(String, true).

Note that UCDParser uses this latter and so it always checks the namespace of
the UCDs to parse. This behaviour can not be changed.