Finalise the repository/library rename operation (from ucdvalidator to ucidy):   - change the package name (ari.ucd -> ari.ucidy)   - adapt the ANT script to generate jars starting with 'ucidy_' instead of     'uv_'   - update the UML diagram (and its file name)
The README has been slightly changed in order to link the new documentation
resources: the UML diagram and the Wiki documentation.

Additionally, the WARNING of UCDWordList.add(UCDWord) (l.141) has been
displayed on stdout instead of stderr, and the WARNING message of
UCDParser.parseWordList(...) (l.523) has been completed with 2 other reasons of
why a UCD word has not been accepted in UCDWordList.