Update official words (without the deprecated words) + Support deprecated words.
This list of official IVOA words is fetched from
http://cdsweb.u-strasbg.fr/UCD/ucd1p-words-20180527.txt.

The list of deprecated words is fetched from
http://cdsweb.u-strasbg.fr/UCD/ucd1p-deprecated.txt.

Deprecated words are used only for better error messages and correction
suggestions. Basically, when a word is not recognised, Ucidy search among the
deprecated words ; if a match is found, the error message reports the problem
as a deprecated word and suggest the appropriate serie of words to replace it.